refactor: domain, view, generator 패키지 분류
