+++ src/main/java/racingcar/domain/LoadMovable.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public class LoadMovable implements Movable {
     private static final int MOVE_CONDITION = 4;
