+++ src/main/java/racingcar/domain/Movable.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public interface Movable {
     boolean move(int num);
