+++ src/test/java/racingcar/CarNameTest.java
@@ -3,6 +3,7 @@ package racingcar;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.domain.CarName;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
